+++ src/test/php/lang/reflect/ReflectionObjectTest.php
@@ -410,10 +410,10 @@ class ReflectionObjectTest extends \PHPUnit_Framework_TestCase
     public function annotationsReturnsListOfAllAnnotation()
     {
         $this->assertEquals(
-                ['SomeAnnotation'    => new Annotation('SomeAnnotation', $this->refClass2->getName()),
-                 'AnotherAnnotation' => new Annotation('AnotherAnnotation', $this->refClass2->getName())
+                [new Annotation('SomeAnnotation', $this->refClass2->getName()),
+                 new Annotation('AnotherAnnotation', $this->refClass2->getName())
                 ],
-                $this->refClass2->annotations()
+                $this->refClass2->annotations()->all()
         );
     }
 }