+++ src/test/php/lang/reflect/ReflectionParameterTest.php
@@ -205,10 +205,10 @@ class ReflectionParameterTest extends \PHPUnit_Framework_TestCase
     public function annotationsReturnsListOfAllAnnotationForFunctionParameter()
     {
         $this->assertEquals(
-                ['ParamAnno'         => new Annotation('ParamAnno', 'stubbles\lang\reflect\test_function()#param'),
-                 'AnotherAnnotation' => new Annotation('AnotherAnnotation', 'stubbles\lang\reflect\test_function()#param')
+                [new Annotation('ParamAnno', 'stubbles\lang\reflect\test_function()#param'),
+                 new Annotation('AnotherAnnotation', 'stubbles\lang\reflect\test_function()#param')
                 ],
-                $this->refParamFunction->annotations('SomeAnnotation')
+                $this->refParamFunction->annotations()->all()
         );
     }
 
@@ -219,10 +219,10 @@ class ReflectionParameterTest extends \PHPUnit_Framework_TestCase
     public function annotationsReturnsListOfAllAnnotationForMethodParameter()
     {
         $this->assertEquals(
-                ['ParamAnno'         => new Annotation('ParamAnno', 'stubbles\lang\reflect\ParamTestHelper::paramTest()#param'),
-                 'AnotherAnnotation' => new Annotation('AnotherAnnotation', 'stubbles\lang\reflect\ParamTestHelper::paramTest()#param')
+                [new Annotation('ParamAnno', 'stubbles\lang\reflect\ParamTestHelper::paramTest()#param'),
+                 new Annotation('AnotherAnnotation', 'stubbles\lang\reflect\ParamTestHelper::paramTest()#param')
                 ],
-                $this->refParamMethod1->annotations()
+                $this->refParamMethod1->annotations()->all()
         );
     }