+++ src/test/php/lang/reflect/ReflectionPropertyTest.php
@@ -233,10 +233,10 @@ class ReflectionPropertyTest extends \PHPUnit_Framework_TestCase
     public function annotationsReturnsListOfAllAnnotation()
     {
         $this->assertEquals(
-                ['SomeAnnotation'    => new Annotation('SomeAnnotation', 'stubbles\lang\reflect\TestProperty1->property'),
-                 'AnotherAnnotation' => new Annotation('AnotherAnnotation', 'stubbles\lang\reflect\TestProperty1->property')
+                [new Annotation('SomeAnnotation', 'stubbles\lang\reflect\TestProperty1->property'),
+                 new Annotation('AnotherAnnotation', 'stubbles\lang\reflect\TestProperty1->property')
                 ],
-                $this->refProperty->annotations()
+                $this->refProperty->annotations()->all()
         );
     }
 }